move helpInformation
