+++ Devscord.DiscordFramework/Services/Models/Description.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Text;
+
+namespace Watchman.DomainModel.Help
+{
+    public class Description
+    {
+        public string Name { get; set; }
+        public string Details { get; set; }
+    }
+}
