+++ Devscord.DiscordFramework/Services/Models/ServerHelpInformation.cs
@@ -1,0 +1,7 @@
+﻿namespace Watchman.DomainModel.Help
+{
+    public class ServerHelpInformation : DefaultHelpInformation
+    {
+        public ulong ServerId { get; set; }
+    }
+}
